requirements.txt
geopy==1.18.1
matplotlib==3.0.3
timestring==1.6.2
numpy==1.16.2
pandas==0.24.1
nltk==3.4
plotly==3.7.0
tabulate==0.8.2
Pillow==5.4.1
python_dateutil==2.8.0
langdetect==1.0.7
wordcloud==1.5.0
